+++ modules/cardsData.js
@@ -1,0 +1,12 @@
+const myPortfolio = [
+  {
+    id: 1,
+    title: 'My Title',
+    briefDes: 'Brief description',
+    techList: ['html', 'bootstrap', 'Ruby'],
+    image: './',
+    detailedDes: 'Detailed des',
+  },
+];
+
+export default myPortfolio;
